requirements.txt
telethon
cryptg
pymongo
motor
telegraph
python-dotenv==0.17.1